misc fixes and refactoring (moved filter_analyzers() to controller.py)
