+++ api_app/analyzers_manager/controller.py
@@ -8,11 +8,11 @@ from django.core.cache import cache
 from intel_owl.settings import CELERY_QUEUES
 from intel_owl.celery import app as celery_app
 
-from ..models import Job
-from ..helpers import generate_sha256, get_now
-from ..exceptions import AlreadyFailedJobException
-from .serializers import AnalyzerConfigSerializer
+from . import serializers
 from .models import AnalyzerReport
+from ..models import Job
+from ..helpers import get_now
+from ..exceptions import AlreadyFailedJobException, NotRunnableAnalyzer
 
 
 logger = logging.getLogger(__name__)
@@ -35,6 +35,93 @@ def build_cache_key(job_id: int):
     return f"job.{job_id}.analyzers_manager.task_ids"
 
 
+def filter_analyzers(
+    serialized_data, analyzers_requested, warnings, run_all=False
+) -> list:
+    cleaned_analyzer_list = []
+
+    analyzers_config = serializers.AnalyzerConfigSerializer.read_and_verify_config()
+
+    if analyzers_requested == "__all__":
+        # select all
+        analyzers_requested = list(analyzers_config.keys())
+
+    for analyzer in analyzers_requested:
+        try:
+            if analyzer not in analyzers_config:
+                raise NotRunnableAnalyzer(f"{analyzer} not available in configuration")
+
+            analyzer_config = analyzers_config[analyzer]
+
+            if serialized_data["is_sample"]:
+                if not analyzer_config.get("type", None) == "file":
+                    raise NotRunnableAnalyzer(
+                        f"{analyzer} won't be run because does not support files."
+                    )
+                if (
+                    analyzer_config.get("supported_filetypes", [])
+                    and serialized_data["file_mimetype"]
+                    not in analyzer_config["supported_filetypes"]
+                ):
+                    raise_message = (
+                        f"{analyzer} won't be run because mimetype."
+                        f"{serialized_data['file_mimetype']} is not supported."
+                        f"Supported are:"
+                        f"{analyzer_config['supported_filetypes']}."
+                    )
+                    raise NotRunnableAnalyzer(raise_message)
+                if (
+                    analyzer_config.get("not_supported_filetypes", "")
+                    and serialized_data["file_mimetype"]
+                    in analyzer_config["not_supported_filetypes"]
+                ):
+                    raise_message = f"""
+                        {analyzer} won't be run because mimetype
+                        {serialized_data['file_mimetype']} is not supported.
+                        Not supported are:{analyzer_config['not_supported_filetypes']}.
+                    """
+                    raise NotRunnableAnalyzer(raise_message)
+            else:
+                if not analyzer_config.get("type", None) == "observable":
+                    raise NotRunnableAnalyzer(
+                        f"{analyzer} won't be run because does not support observable."
+                    )
+                if serialized_data[
+                    "observable_classification"
+                ] not in analyzer_config.get("observable_supported", []):
+                    raise NotRunnableAnalyzer(
+                        f"""
+                        {analyzer} won't be run because does not support
+                         observable type {serialized_data['observable_classification']}.
+                        """
+                    )
+            if analyzer_config.get("disabled", False):
+                raise NotRunnableAnalyzer(f"{analyzer} is disabled, won't be run.")
+            if serialized_data["force_privacy"] and analyzer_config.get(
+                "leaks_info", False
+            ):
+                raise NotRunnableAnalyzer(
+                    f"{analyzer} won't be run because it leaks info externally."
+                )
+            if serialized_data["disable_external_analyzers"] and analyzer_config.get(
+                "external_service", False
+            ):
+                raise NotRunnableAnalyzer(
+                    f"{analyzer} won't be run because you filtered external analyzers."
+                )
+        except NotRunnableAnalyzer as e:
+            if run_all:
+                # in this case, they are not warnings but expected and wanted behavior
+                logger.debug(e)
+            else:
+                logger.warning(e)
+                warnings.append(str(e))
+        else:
+            cleaned_analyzer_list.append(analyzer)
+
+    return cleaned_analyzer_list
+
+
 def start_analyzers(
     job_id: int,
     analyzers_to_execute: list,
@@ -45,7 +132,7 @@ def start_analyzers(
     analyzer_task_id_map = {}
 
     # get analyzer config
-    analyzers_config = AnalyzerConfigSerializer.read_and_verify_config()
+    analyzers_config = serializers.AnalyzerConfigSerializer.read_and_verify_config()
 
     # get job
     job = Job.objects.get(pk=job_id)
