+++ api_app/analyzers_manager/serializers.py
@@ -3,8 +3,10 @@
 from enum import Enum
 
 from rest_framework import serializers as rfs
+from django.utils.module_loading import import_string
 
 from api_app.core.serializers import AbstractConfigSerializer
+from .controller import build_import_path
 from .models import AnalyzerReport
 
 
@@ -51,30 +53,29 @@ class AnalyzerConfigSerializer(AbstractConfigSerializer):
         GENERIC = "generic"
 
     # Required fields
-    type = rfs.ChoiceField(required=True, choices=[c.name for c in TypeChoices])
+    type = rfs.ChoiceField(required=True, choices=[c.value for c in TypeChoices])
     external_service = rfs.BooleanField(required=True)
     # Optional Fields
     leaks_info = rfs.BooleanField(required=False, default=False)
     run_hash = rfs.BooleanField(required=False, default=False)
     run_hash_type = rfs.ChoiceField(
-        required=False, choices=[c.name for c in HashChoices]
+        required=False, choices=[c.value for c in HashChoices]
     )
     supported_filetypes = rfs.ListField(required=False)
     not_supported_filetypes = rfs.ListField(required=False)
     observable_supported = rfs.MultipleChoiceField(
-        choices=[c.name for c in ObservableTypes]
+        choices=[c.value for c in ObservableTypes], required=False
     )
 
     def validate_python_module(self, python_module: str):
-        from .general import build_import_path
-        from django.utils.module_loading import import_string
-
         clspath = build_import_path(
             python_module,
             observable_analyzer=(
-                self.data["type"] == self.TypeChoices.OBSERVABLE
-                or self.data["type"] == self.TypeChoices.FILE
-                and self.data["run_hash"]
+                self.initial_data["type"] == self.TypeChoices.OBSERVABLE.value
+                or (
+                    self.initial_data["type"] == self.TypeChoices.FILE.value
+                    and self.initial_data.get("run_hash", False)
+                )
             ),
         )
         try:
