+++ api_app/core/serializers.py
@@ -148,8 +148,8 @@ class AbstractConfigSerializer(rfs.Serializer):
         must be called after `.is_valid()`.
         """
         assert (
-            self._is_valid_flag == True
-        ), "Cannot call `._read_secrets()` before `.is_valid()` or if validation failed."
+            self._is_valid_flag
+        ), "Cannot call `._read_secrets()` before `.is_valid()` or validation failed."
 
         secrets = {}
         for key_name, secret_dict in self.data["secrets"].items():
