+++ api_app/helpers.py
@@ -10,9 +10,7 @@ from magic import from_buffer as magic_from_buffer
 
 from django.utils import timezone
 
-from .exceptions import NotRunnableAnalyzer
-from api_app import models
-from api_app.analyzers_manager.serializers import AnalyzerConfigSerializer
+from . import models
 
 logger = logging.getLogger(__name__)
 
@@ -47,93 +45,6 @@ def calculate_mimetype(file_buffer, file_name):
             calculated_mimetype = "application/x-dex"
 
     return calculated_mimetype
-
-
-def filter_analyzers(
-    serialized_data, analyzers_requested, warnings, run_all=False
-) -> list:
-    cleaned_analyzer_list = []
-
-    analyzers_config = AnalyzerConfigSerializer.read_and_verify_config()
-
-    if analyzers_requested == "__all__":
-        # select all
-        analyzers_requested = list(analyzers_config.keys())
-
-    for analyzer in analyzers_requested:
-        try:
-            if analyzer not in analyzers_config:
-                raise NotRunnableAnalyzer(f"{analyzer} not available in configuration")
-
-            analyzer_config = analyzers_config[analyzer]
-
-            if serialized_data["is_sample"]:
-                if not analyzer_config.get("type", None) == "file":
-                    raise NotRunnableAnalyzer(
-                        f"{analyzer} won't be run because does not support files."
-                    )
-                if (
-                    analyzer_config.get("supported_filetypes", [])
-                    and serialized_data["file_mimetype"]
-                    not in analyzer_config["supported_filetypes"]
-                ):
-                    raise_message = (
-                        f"{analyzer} won't be run because mimetype."
-                        f"{serialized_data['file_mimetype']} is not supported."
-                        f"Supported are:"
-                        f"{analyzer_config['supported_filetypes']}."
-                    )
-                    raise NotRunnableAnalyzer(raise_message)
-                if (
-                    analyzer_config.get("not_supported_filetypes", "")
-                    and serialized_data["file_mimetype"]
-                    in analyzer_config["not_supported_filetypes"]
-                ):
-                    raise_message = f"""
-                        {analyzer} won't be run because mimetype
-                        {serialized_data['file_mimetype']} is not supported.
-                        Not supported are:{analyzer_config['not_supported_filetypes']}.
-                    """
-                    raise NotRunnableAnalyzer(raise_message)
-            else:
-                if not analyzer_config.get("type", None) == "observable":
-                    raise NotRunnableAnalyzer(
-                        f"{analyzer} won't be run because does not support observable."
-                    )
-                if serialized_data[
-                    "observable_classification"
-                ] not in analyzer_config.get("observable_supported", []):
-                    raise NotRunnableAnalyzer(
-                        f"""
-                        {analyzer} won't be run because does not support
-                         observable type {serialized_data['observable_classification']}.
-                        """
-                    )
-            if analyzer_config.get("disabled", False):
-                raise NotRunnableAnalyzer(f"{analyzer} is disabled, won't be run.")
-            if serialized_data["force_privacy"] and analyzer_config.get(
-                "leaks_info", False
-            ):
-                raise NotRunnableAnalyzer(
-                    f"{analyzer} won't be run because it leaks info externally."
-                )
-            if serialized_data["disable_external_analyzers"] and analyzer_config.get(
-                "external_service", False
-            ):
-                raise NotRunnableAnalyzer(
-                    f"{analyzer} won't be run because you filtered external analyzers."
-                )
-        except NotRunnableAnalyzer as e:
-            if run_all:
-                # in this case, they are not warnings but expected and wanted behavior
-                logger.debug(e)
-            else:
-                logger.warning(e)
-                warnings.append(str(e))
-        else:
-            cleaned_analyzer_list.append(analyzer)
-
-    return cleaned_analyzer_list
 
 
 def get_binary(job_id, job_object=None):
