+++ intel_owl/tasks.py
@@ -2,7 +2,6 @@
 # See the file 'LICENSE' for copying permission.
 
 from __future__ import absolute_import, unicode_literals
-import importlib
 from celery import shared_task
 from guardian import utils
